fix(db): customer_id 过滤参数显式 ::bigint,修合成 ID 客户全线 50000
PG 对 ($n = 0 OR customer_id = $n) 会从 "= 0" 把参数推断为 int4,
门户注册发的隔离空间合成 ID(9e9 段)直接 22003 溢出,新注册用户的
/orders /bills 等端点全部 50000(真机首现「订单加载失败,请稍后重试」)。
对全部 customer_id 条件加 ::bigint 显式转型(bill_id/offer_id 等
真实 FK 不会超 int4 不动);respondErr 兜底分支补错误日志,500 不再无声。

+++ internal/domain/billing/pg.go
@@ -34,7 +34,7 @@ const billCols = `id, bill_no, customer_id, customer_name, legal_entity_id, lega
 // ListBills 列出账单;customerID=0 返回全部,否则按客户过滤。
 func (s *PGStore) ListBills(ctx context.Context, customerID int64) ([]Bill, error) {
 	rows, err := s.db.Query(ctx,
-		`SELECT `+billCols+` FROM bills WHERE ($1 = 0 OR customer_id = $1) ORDER BY id`, customerID)
+		`SELECT `+billCols+` FROM bills WHERE ($1::bigint = 0 OR customer_id = $1) ORDER BY id`, customerID)
 	if err != nil {
 		return nil, fmt.Errorf("billing: list bills: %w", err)
 	}
